requirements.txt
Django>=4.2.16,<4.3
djangorestframework>=3.14.0,<3.15
pinyin>=0.4.0 